Things should be working fine

+++ DBS/src/backup/BackupSend.java
@@ -11,8 +11,10 @@ public class BackupSend extends Thread {
     private Integer time;
     private Integer storeds;
     private Integer tries;
+    private Integer chunkN;
+    private Boolean isFile;
 
-    public BackupSend(String filePath, Integer repDegree) {
+    public BackupSend(String filePath, Integer repDegree, Boolean file, Integer newChunkNo) {
 
         fm = new FileManager(filePath, repDegree, false);
         fileHash = fm.getHashString().toString();
@@ -20,53 +22,91 @@ public class BackupSend extends Thread {
         time = 500;
         storeds = 0;
         tries = 0;
+        isFile = file;
+        chunkN = newChunkNo;
     }
 
     public void run() {
 
-        Boolean enoughSpace;
-        enoughSpace = fm.split();
+        if (isFile) {
+            Boolean enoughSpace;
+            enoughSpace = fm.split();
 
-        if (enoughSpace) {
-            Integer chunkNo = 0;
+            if (enoughSpace) {
+                Integer chunkNo = 0;
 
-            while (fm.readChunk(chunkNo)) {
+                while (fm.readChunk(chunkNo)) {
 
-                String message = "";
+                    String message = "";
                     message = "PUTCHUNK " + Main.getVersion() + " " + fileHash + " " + chunkNo + " " + fm.getRep() +
                             Main.getCRLF() + Main.getCRLF();
 
-                byte[] mssg = message.getBytes(StandardCharsets.ISO_8859_1);
-                byte[] mssg1 = Main.appendArray(mssg, fm.getChunkData());
+                    byte[] mssg = message.getBytes(StandardCharsets.ISO_8859_1);
+                    byte[] mssg1 = Main.appendArray(mssg, fm.getChunkData());
 
-                while (storeds < fm.getRep() && tries < 5) {
+                    while (storeds < fm.getRep() && tries < 5) {
 
-                    storeds = 0;
-                    Main.getBackup().send(mssg1);
+                        storeds = 0;
+                        Main.getBackup().send(mssg1);
 
+                        try {
+                            sleep(time);
+                        } catch (InterruptedException e) {
+                            e.printStackTrace();
+                        }
 
-                    try {
-                        sleep(time);
-                    } catch (InterruptedException e) {
-                        e.printStackTrace();
+                        tries++;
+                        time += time;
                     }
 
-                    tries++;
-                    time += time;
+                    fm.deleteChunk(chunkNo);
+
+                    chunkNo++;
+                    storeds = 0;
+                    time = 500;
+                    tries = 0;
                 }
-
-                fm.deleteChunk(chunkNo);
-
-                chunkNo++;
-                storeds = 0;
-                time = 500;
-                tries = 0;
+            } else {
+                System.out.println("Not enough space for backup!");
             }
+
+            System.out.println("Finished backup of fileid " + fileHash);
         } else {
-            System.out.println("Not enough space for backup!");
+            backupChunk(chunkN);
+        }
+    }
+
+    public void backupChunk(Integer chunkNo) {
+        fm.readChunk(chunkNo);
+        String message = "";
+        message = "PUTCHUNK " + Main.getVersion() + " " + fileHash + " " + chunkNo + " " + fm.getRep() +
+                Main.getCRLF() + Main.getCRLF();
+
+        byte[] mssg = message.getBytes(StandardCharsets.ISO_8859_1);
+        byte[] mssg1 = Main.appendArray(mssg, fm.getChunkData());
+
+        while (storeds < fm.getRep() && tries < 5) {
+
+            storeds = 0;
+            Main.getBackup().send(mssg1);
+
+
+            try {
+                sleep(time);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+
+            tries++;
+            time += time;
         }
 
-        System.out.println("Finished backup of fileid " + fileHash);
+        fm.deleteChunk(chunkNo);
+
+        chunkNo++;
+        storeds = 0;
+        time = 500;
+        tries = 0;
     }
 
     public synchronized void incStoreds(Integer newreps) {
